@@ -15,19 +15,22 @@ public class Feed implements CommandExecutor {
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         YamlConfiguration config = ConfigHandler.getConfig();
         String prefix = ChatColor.translateAlternateColorCodes('&', config.getString("Prefix"));
-        if (args.length > 0 && sender.hasPermission("usefulcommands.feed.others")) {
-            try {
-                feed(Bukkit.getPlayer(args[0]));
-                sender.sendMessage(prefix + Bukkit.getPlayer(args[0]).getName() + " has been fed");
-            } catch (Exception e) {
-                sender.sendMessage(prefix + args[0] + " is not a valid player");
+        if (args.length > 0) {
+            if (!sender.hasPermission("usefulcommands.feed.others")) {
+                sender.sendMessage(prefix + "§cI'm sorry but you do not have permission to perform this command. Please contact the server administrator if you believe that this is in error.");
+                return true;
             }
-        } else if (sender instanceof Player && args.length == 0) {
-            feed((Player) sender);
-        } else if (!(sender instanceof Player)) {
-            sender.sendMessage(prefix + "You must be a player to use this command");
+            Player target = Bukkit.getPlayer(args[0]);
+            if (target != null) {
+                feed(target);
+                sender.sendMessage(prefix + "You have fed " + target.getName() + ".");
+            } else {
+                sender.sendMessage(prefix + "That player is not online.");
+            }
+        } else if (sender instanceof Player player) {
+            feed(player);
         } else {
-            sender.sendMessage("§cI'm sorry but you do not have permission to perform this command. Please contact the server administrator if you believe that this is in error.");
+            sender.sendMessage(prefix + "You must be a player to use this command");
         }
         return true;
     }
