@@ -14,8 +14,12 @@ public class Suicide implements CommandExecutor {
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         YamlConfiguration config = ConfigHandler.getConfig();
         String prefix = ChatColor.translateAlternateColorCodes('&', config.getString("Prefix"));
-        ((Player) sender).setHealth(0);
-        sender.sendMessage(prefix + "You have died.");
+        if (sender instanceof Player player) {
+            player.setHealth(0);
+            sender.sendMessage(prefix + "You have died.");
+        } else {
+            sender.sendMessage(prefix + "bro how tf is console going to kill itself");
+        }
         return true;
     }
 }
